+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Gluecodium project Release Notes
 
+## Unreleased
+### Features:
+  * Added support for `Locale` type literal values.
+
 ## 11.0.1
 Release date: 2022-02-22
 ### Bug fixes: